Move TopGraphObserver to `shared-ui`.
